Faz o tratamento com exceções

+++ src/main/java/br/com/ferdbgg/springteste/services/UsuarioService.java
@@ -4,10 +4,14 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import br.com.ferdbgg.springteste.entities.Usuario;
 import br.com.ferdbgg.springteste.repositories.UsuarioRepository;
+import br.com.ferdbgg.springteste.services.exceptions.BancoDadosException;
+import br.com.ferdbgg.springteste.services.exceptions.ObjetoNaoEncontradoException;
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class UsuarioService {
@@ -21,7 +25,7 @@ public class UsuarioService {
 
     public Usuario findById(Integer id) {
         Optional<Usuario> usuario = this.usuarioRepository.findById(id);
-        return usuario.isPresent() ? usuario.get() : new Usuario();
+        return usuario.orElseThrow(() -> new ObjetoNaoEncontradoException(id));
     }
 
     public Usuario insert(Usuario novoUsuario) {
@@ -29,17 +33,31 @@ public class UsuarioService {
     }
 
     public void deleteById(Integer id) {
-        this.usuarioRepository.deleteById(id);
+        Optional<Usuario> usuario = this.usuarioRepository.findById(id);
+        if(usuario.isPresent()) {
+            try {
+                this.usuarioRepository.deleteById(id);
+            } catch(DataIntegrityViolationException exception) {
+                throw new BancoDadosException(exception.getMessage());
+            }
+        } else {
+            throw new ObjetoNaoEncontradoException(id);
+        }
     }
 
     public Usuario update(Integer id, Usuario usuarioNovosDados) {
         Usuario usuarioMonitoradoJPA = this.usuarioRepository.getReferenceById(id);
-        this.atualizarDados(usuarioMonitoradoJPA, usuarioNovosDados);
-        return this.usuarioRepository.save(usuarioMonitoradoJPA);
+        //Dá EntityNotFoundException ao tentar acessar uma entidade que não está persistida
+        try {
+            this.atualizarDados(usuarioMonitoradoJPA, usuarioNovosDados);
+            return this.usuarioRepository.save(usuarioMonitoradoJPA);
+        } catch(EntityNotFoundException exception) {
+            throw new ObjetoNaoEncontradoException(id);
+        }
     }
 
     private void atualizarDados(Usuario usuarioMonitoradoJPA, Usuario usuarioNovosDados) {
-        //TODO: Verificar se a regra de negócio permite colocar null ou não
+        //Verificar se a regra de negócio permite colocar null ou não
         if(usuarioNovosDados.getNome() != null) {
             usuarioMonitoradoJPA.setNome(usuarioNovosDados.getNome()); 
         }
